test : 즐겨찾기 단위 테스트 구현
