--- src/test/java/nextstep/subway/unit/FavoriteTest.java
@@ -1,37 +1,0 @@
-package nextstep.subway.unit;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-public class FavoriteTest {
-
-	@BeforeEach
-	void setUp() {
-
-	}
-
-	@Test
-	@DisplayName("즐겨찾기 등록")
-	void addFavorite() {
-
-	}
-
-	@Test
-	@DisplayName("즐겨찾기 삭제")
-	void removeFavorite() {
-
-	}
-
-	@Test
-	@DisplayName("동일 즐겨찾기 등록")
-	void addSameFavorite() {
-
-	}
-
-	@Test
-	@DisplayName("등록되지 않은 즐겨찾기 삭제")
-	void removeNotExistingFavorite() {
-
-	}
-}
